refactor(code-editor): reorganize panel testing utilities and update import paths
- Moved panel-specific testing utilities to a new file, enhancing organization and maintainability.
- Updated import paths in the Panel test file to reflect the new location of the testing utilities.
- Improved clarity and accessibility of testing functions for the Panel component.
